added loader and update flow with verify token

+++ src/utils/helper.js
@@ -12,27 +12,47 @@ export const checkAuthAndRedirect = async (navigate,from) => {
 
     const domainFromUrl = window.location.hostname;
     const expectedDomain = process.env.REACT_APP_DOMAIN;
+    if(from == 'Register'){
+        if (domainFromUrl != expectedDomain) {
+            window.location.href = `http://${process.env.REACT_APP_DOMAIN}/`;
+        }
+    }
 
-    if (domainFromUrl != expectedDomain) {
-        var token = localStorage.getItem(domainFromUrl);
-        if (token && token.trim() !== '') {
-            // Redirect to dashboard if token is present
-            // navigate('/dashboard');
-            var response = await verifyToken();
-            console.log('response :',response);
-            if (response) {
-                navigate('/dashboard');
-            } else {
+    if(from == 'DomainName'){
+        if (domainFromUrl != expectedDomain) {
+            var token = localStorage.getItem(domainFromUrl);
+            if (token && token.trim() !== '') {
+                var response = await verifyToken();
+                if (response) {
+                    navigate('/dashboard');
+                } else {
+                    localStorage.clear();
+                    navigate('/signin');
+                }
+            }else{
                 localStorage.clear();
                 navigate('/signin');
             }
-        }else{
-            console.log('3333');
-            localStorage.clear();
-            navigate('/signin');
         }
-    }else{
-        if(from == "Signin"){
+    }
+    if(from == 'Signin'){
+        if (domainFromUrl != expectedDomain) {
+            var token = localStorage.getItem(domainFromUrl);
+            if (token && token.trim() !== '') {
+                // Redirect to dashboard if token is present
+                // navigate('/dashboard');
+                var response = await verifyToken();
+                console.log('response :',response);
+                if (response) {
+                    navigate('/dashboard');
+                } else {
+                    localStorage.clear();
+                }
+            }else{
+                console.log('3333');
+                localStorage.clear();
+            }
+        }else{
             navigate('/domain-signin');
         }
     }
